scGenePT/requirements_conda.txt
absl-py==2.1.0
accelerate==1.0.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
alembic==1.13.3
aniso8601==9.0.1
anndata==0.10.9
array_api_compat==1.8
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cached-property==1.5.2
cachetools==5.5.0
cell-gears==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.33.0
datasets==2.21.0
dcor==0.6
debugpy==1.8.8
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
docker==7.1.0
docrep==0.3.2
einops==0.8.0
et-xmlfile==1.1.0
etils==1.9.4
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
Flask==3.0.3
flax==0.9.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
h5py==3.12.1
huggingface-hub==0.25.1
humanize==4.10.0
idna==3.10
igraph==0.11.6
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jax==0.4.33
jaxlib==0.4.33
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
legacy-api-wrap==1.4
leidenalg==0.10.2
lightning-utilities==0.11.7
llvmlite==0.43.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_collections==0.1.1
ml_dtypes==0.5.0
mlflow==2.16.2
mlflow-skinny==2.16.2
mpmath==1.3.0
msgpack==1.1.0
mudata==0.3.1
multidict==6.1.0
multipledispatch==1.0.0
multiprocess==0.70.16
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
numpyro==0.15.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt_einsum==3.4.0
optax==0.2.3
orbax==0.1.7
orbax-checkpoint==0.6.4
packaging==24.1
pandas==2.2.3
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydot==3.0.2
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.4
pyro-api==0.1.2
pyro-ppl==1.9.1
pytest==8.3.3
python-dateutil==2.9.0.post0
pytorch-lightning==1.9.5
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rsa==4.9
safetensors==0.4.5
scanpy==1.10.3
flash_attn==1.0.4
scgpt==0.2.1
scib==1.1.5
scikit-learn==1.5.2
scikit-misc==0.5.1
scipy==1.14.1
scvi-tools==0.20.3
seaborn==0.13.2
session_info==1.0.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
stack-data==0.6.3
statsmodels==0.14.3
stdlib-list==0.10.0
sympy==1.13.3
tenacity==9.0.0
tensorstore==0.1.66
texttable==1.7.0
threadpoolctl==3.5.0
tokenizers==0.20.0
tomli==2.1.0
toolz==0.12.1
torch==2.1.2
torch_geometric==2.5.3
torchdata==0.7.1
torchmetrics==1.4.2
torchtext==0.16.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.6
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.2
